ref: - move some functions into task_generator module which they are related to that functionality - isolate functions preprocessing input goals into separate module "preprocess"
