+++ src/models/goal/impls.rs
@@ -1,14 +1,7 @@
+use super::{Day, Goal, TimeFilter};
+use crate::models::repetition::Repetition;
 use chrono::NaiveDateTime;
 use log::info;
-
-use crate::models::{
-    goal::Tag,
-    repetition::Repetition,
-    slots_iterator::TimeSlotsIterator,
-    task::{NewTask, Task, TaskStatus},
-};
-
-use super::{Day, Goal, TimeFilter};
 
 impl From<String> for Day {
     fn from(day: String) -> Self {
@@ -75,70 +68,6 @@ impl Goal {
     pub fn deadline(mut self, deadline: NaiveDateTime) -> Self {
         self.deadline = Some(deadline);
         self
-    }
-
-    /// Generates a Task/Increment from a Processed Goal
-    /// **Caution!:*** This can only be done after the Goals have been pre-processed!
-    /// Creates and splits the Goal Timeline into one or more segments, making a Task/Increment for each.
-    /// Depending on the Goal Tag, Task/Increments will also get Tags to help with scheduling order:
-    /// - Optional Tag // Todo! add Regular Tag to simplify?
-    /// - Filler Tag
-    /// - FlexDur Tag
-    /// - FlexNum Tag
-    /// - Budget Tag
-    pub fn generate_tasks(
-        self,
-        calendar_start: NaiveDateTime,
-        calendar_end: NaiveDateTime,
-        counter: &mut usize,
-    ) -> Vec<Task> {
-        let mut tasks: Vec<Task> = Vec::new();
-        if self.tags.contains(&Tag::IgnoreForTaskGeneration) {
-            return tasks;
-        }
-
-        if self.tags.contains(&Tag::Budget) {
-            return tasks;
-        }
-        let start = self.start.unwrap_or(calendar_start);
-        let deadline = self.deadline.unwrap_or(calendar_end);
-
-        let time_slots_iterator = TimeSlotsIterator::new(
-            start,
-            deadline,
-            self.repeat,
-            self.filters.clone(),
-            // Todo! add self.before_time filter
-        );
-        dbg!(&time_slots_iterator);
-
-        for timeline in time_slots_iterator {
-            dbg!(&timeline);
-            let task_id = *counter;
-            *counter += 1;
-            // TODO 2023-05-06  | apply Task::new(...)
-            if !timeline.slots.is_empty() && self.min_duration.is_some() {
-                let title = self.title.clone();
-                let duration = self.min_duration.unwrap();
-
-                let new_task = NewTask {
-                    task_id,
-                    title,
-                    duration,
-                    goal: self.clone(),
-                    timeline,
-                    status: TaskStatus::ReadyToSchedule,
-                    timeframe: None,
-                };
-
-                let task = Task::new(new_task);
-
-                dbg!(&task);
-                tasks.push(task);
-            }
-        }
-        dbg!(&tasks);
-        tasks
     }
 }
