+++ src/services/task_generator.rs
@@ -1,241 +1,92 @@
+use crate::models::goal::{Goal, Tag};
+use crate::models::slots_iterator::TimeSlotsIterator;
+use crate::models::task::{NewTask, Task, TaskStatus};
 use chrono::NaiveDateTime;
-use std::collections::BTreeMap;
 
-use crate::models::budget::TaskBudgets;
-use crate::models::goal::{Goal, GoalsMap, Tag};
-use crate::models::input::{Input, TasksToPlace};
-use crate::models::repetition::Repetition;
-use crate::models::task::Task;
+impl Task {
+    /// Create new task
+    pub fn new(new_task: NewTask) -> Task {
+        let start = new_task.timeframe.map(|time| time.start);
+        let deadline = new_task.timeframe.map(|time| time.end);
 
-// Todo 2023-05-05  | Move preprocessing Goals into separate module(s) - generating Tasks then becomes simple
-/// Preprocesses the hierarchy of goals, then for each Goal call Goal.generate_tasks
-/// Preprocessing involves a number of steps:
-/// - add_start_and_end_where_none
-/// - add_filler_goals
-/// - add_optional_flex_duration_regular_goals
-/// - add_optional_flex_number_and_duration_habits_goals
-/// - create min and max budgets (task_budgets.create_task_budgets_config)
-/// - task_budgets.generate_budget_min_and_max_tasks
-pub fn generate_tasks_to_place(input: Input) -> TasksToPlace {
-    let calendar_start = input.calendar_start;
-    let calendar_end = input.calendar_end;
+        Task {
+            id: new_task.task_id,
+            goal_id: new_task.goal.id,
+            title: new_task.title,
+            duration: new_task.duration,
+            status: new_task.status,
+            flexibility: 0,
+            start,
+            deadline,
+            slots: new_task.timeline.slots.into_iter().collect(),
+            tags: new_task.goal.tags,
+            after_goals: new_task.goal.after_goals,
+        }
+    }
+}
 
-    let mut goals = manipulate_input_goals(input);
+impl Goal {
+    /// Generates a Task/Increment from a Processed Goal
+    /// **Caution!:*** This can only be done after the Goals have been pre-processed!
+    /// Creates and splits the Goal Timeline into one or more segments, making a Task/Increment for each.
+    /// Depending on the Goal Tag, Task/Increments will also get Tags to help with scheduling order:
+    /// - Optional Tag // Todo! add Regular Tag to simplify?
+    /// - Filler Tag
+    /// - FlexDur Tag
+    /// - FlexNum Tag
+    /// - Budget Tag
+    pub fn generate_tasks(
+        self,
+        calendar_start: NaiveDateTime,
+        calendar_end: NaiveDateTime,
+        counter: &mut usize,
+    ) -> Vec<Task> {
+        let mut tasks: Vec<Task> = Vec::new();
+        if self.tags.contains(&Tag::IgnoreForTaskGeneration) {
+            return tasks;
+        }
 
-    let mut task_budgets = TaskBudgets::new(&calendar_start, &calendar_end);
-    task_budgets.configure_budgets(&mut goals);
+        if self.tags.contains(&Tag::Budget) {
+            return tasks;
+        }
+        let start = self.start.unwrap_or(calendar_start);
+        let deadline = self.deadline.unwrap_or(calendar_end);
 
-    let mut counter: usize = 0;
-    let mut tasks: Vec<Task> =
-        task_budgets.generate_budget_min_and_max_tasks(&mut goals, &mut counter);
+        let time_slots_iterator = TimeSlotsIterator::new(
+            start,
+            deadline,
+            self.repeat,
+            self.filters.clone(),
+            // Todo! add self.before_time filter
+        );
+        dbg!(&time_slots_iterator);
 
-    for goal in goals {
-        //for regular, filler, optional flexduration regular, optional flexnumber and/or flexduration habit goals
-        let tasks_for_goal: Vec<Task> =
-            goal.1
-                .generate_tasks(calendar_start, calendar_end, &mut counter);
-        tasks.extend(tasks_for_goal);
+        for timeline in time_slots_iterator {
+            dbg!(&timeline);
+            let task_id = *counter;
+            *counter += 1;
+            // TODO 2023-05-06  | apply Task::new(...)
+            if !timeline.slots.is_empty() && self.min_duration.is_some() {
+                let title = self.title.clone();
+                let duration = self.min_duration.unwrap();
+
+                let new_task = NewTask {
+                    task_id,
+                    title,
+                    duration,
+                    goal: self.clone(),
+                    timeline,
+                    status: TaskStatus::ReadyToSchedule,
+                    timeframe: None,
+                };
+
+                let task = Task::new(new_task);
+
+                dbg!(&task);
+                tasks.push(task);
+            }
+        }
         dbg!(&tasks);
-    }
-
-    dbg!(&tasks);
-
-    TasksToPlace {
-        calendar_start,
-        calendar_end,
-        tasks,
-        task_budgets,
-    }
-}
-
-/// Manipulate input which contains goals with start and end dates.
-/// Returns list of Goals after manipulation
-fn manipulate_input_goals(input: Input) -> GoalsMap {
-    let calendar_start = input.calendar_start;
-    let calendar_end = input.calendar_end;
-    let goals = input.goals;
-
-    let mut goals: GoalsMap = populate_goal_dates(goals, calendar_start, calendar_end);
-
-    add_filler_goals(&mut goals);
-
-    // TODO 2023-06-17: removed empty function until need it and develop it add_optional_flex_duration_regular_goals(&mut goals);
-
-    generate_flex_weekly_goals(&mut goals);
-
-    goals
-}
-
-fn populate_goal_dates(
-    mut goals: GoalsMap,
-    calendar_start: NaiveDateTime,
-    calendar_end: NaiveDateTime,
-) -> GoalsMap {
-    for goal in goals.iter_mut() {
-        goal.1.start.get_or_insert(calendar_start);
-        goal.1.deadline.get_or_insert(calendar_end);
-    }
-    goals
-}
-
-/// Generate new goals based on given goals' FlexWeekly repetition
-/// - Note: this function generating goals for goals with FlexWeekly repetition only
-fn generate_flex_weekly_goals(goals: &mut GoalsMap) {
-    let mut generated_goals: GoalsMap = BTreeMap::new();
-    for (goal_id, goal) in goals.iter_mut() {
-        if let Some(Repetition::FlexWeekly(min, max)) = goal.repeat {
-            //Flex repeat goals are handled as follows:
-            //If given a goal with 3-5x/week, create 3 goals and 2 extra optional goals
-            goal.repeat = Some(Repetition::Weekly(1));
-
-            // Create repeated goals and optional repeated goals
-            for number in 1..max {
-                // 1.. because we're leaving the initial goal
-                let mut template_goal = goal.clone();
-                template_goal.id.push_str("-repeat-");
-
-                if number < min {
-                    // Repeated goal
-                    template_goal.id.push_str(&number.to_string());
-                    generated_goals.insert(template_goal.id.clone(), template_goal);
-                } else {
-                    // Optional repeated goal
-                    template_goal.id.push_str("opt-");
-                    template_goal.id.push_str(&number.to_string());
-                    template_goal.tags.push(Tag::Optional);
-                    generated_goals.insert(template_goal.id.clone(), template_goal);
-                }
-            }
-            generated_goals.insert(goal_id.to_owned(), goal.to_owned());
-        }
-    }
-
-    goals.extend(generated_goals);
-}
-
-pub fn add_filler_goals(goals: &mut GoalsMap) {
-    let mut results: GoalsMap = BTreeMap::new();
-    let mut ignore: Vec<String> = Vec::new();
-    let mut children_to_add: Vec<(String, String)> = Vec::new();
-    for goal in goals.iter() {
-        if goal.1.children.is_some() && goal.1.budgets.is_none() {
-            let mut duration_of_children: usize = 0;
-            for child in goal.1.children.clone().unwrap().iter() {
-                let child_goal = goals.get(child).unwrap();
-                duration_of_children += child_goal.min_duration.unwrap();
-            }
-            let difference = goal.1.min_duration.unwrap() - duration_of_children;
-            if difference > 0 {
-                let mut filler_goal = goal.1.clone();
-                children_to_add.push((goal.1.id.clone(), filler_goal.id.clone()));
-                filler_goal.title.push_str(" filler");
-                filler_goal.min_duration = Some(difference);
-                filler_goal.tags.push(Tag::Filler);
-                results.insert(filler_goal.id.clone(), filler_goal);
-                ignore.push(goal.1.id.clone());
-            }
-        }
-    }
-    for goal_id_to_ignore in ignore {
-        goals
-            .get_mut(&goal_id_to_ignore)
-            .unwrap()
-            .tags
-            .push(Tag::IgnoreForTaskGeneration);
-    }
-    for parent_child in children_to_add {
-        goals
-            .get_mut(&parent_child.0)
-            .unwrap()
-            .children
-            .as_mut()
-            .unwrap()
-            .push(parent_child.1.clone());
-    }
-    goals.extend(results);
-}
-
-#[allow(dead_code)]
-fn get_1_hr_goals(goal: Goal) -> Vec<Goal> {
-    let mut goals = vec![];
-    let dur = goal.min_duration.unwrap();
-    for _ in 0..dur {
-        let mut g = goal.clone();
-        g.min_duration = Some(1);
-        goals.push(g);
-    }
-    goals
-}
-
-// test
-#[cfg(test)]
-mod tests {
-    mod generate_flex_weekly_goals {
-        use std::collections::BTreeMap;
-
-        use chrono::Duration;
-
-        use crate::{
-            models::{
-                goal::{Goal, GoalsMap, Tag},
-                repetition::Repetition,
-                slot::Slot,
-            },
-            services::task_generator::generate_flex_weekly_goals,
-        };
-
-        /// Test generating flex weekly goals based on one task with 1-3/week
-        /// ```markdown
-        /// Input:
-        ///     Goal:
-        ///         title: side project
-        ///         min_duration: 8
-        ///         repeat: 1-3/week
-        ///
-        /// Output:
-        ///     Goal:
-        ///         id: 1-repeat-opt-1
-        ///         title: side project
-        ///         min_duration: 8
-        ///         repeat: 1/week
-        ///     Goal:
-        ///         id: 1-repeat-opt-2
-        ///         title: side project
-        ///         min_duration: 8
-        ///         repeat: 1/week
-        ///
-        /// ```
-        #[test]
-        fn test_single_goal() {
-            let goal_dates = Slot::mock(Duration::days(31), 2022, 10, 1, 0, 0);
-
-            let mut input_goal = Goal::mock("1", "side project", goal_dates);
-            input_goal.min_duration = Some(8);
-            input_goal.repeat = Some(Repetition::FlexWeekly(1, 3));
-
-            let mut input_goals: GoalsMap = BTreeMap::new();
-            input_goals.insert(input_goal.id.clone(), input_goal);
-
-            generate_flex_weekly_goals(&mut input_goals);
-
-            let mut expected_goal_1 = Goal::mock("1", "side project", goal_dates);
-            expected_goal_1.min_duration = Some(8);
-            expected_goal_1.repeat = Some(Repetition::Weekly(1));
-
-            let mut expected_goal_2 = expected_goal_1.clone();
-            expected_goal_2.id = "1-repeat-opt-1".to_string();
-            expected_goal_2.tags.push(Tag::Optional);
-
-            let expected_goal_3 = expected_goal_2.clone();
-            expected_goal_2.id = "1-repeat-opt-2".to_string();
-
-            let mut expected_goals = GoalsMap::new();
-            expected_goals.insert(expected_goal_1.id.clone(), expected_goal_1);
-            expected_goals.insert(expected_goal_2.id.clone(), expected_goal_2);
-            expected_goals.insert(expected_goal_3.id.clone(), expected_goal_3);
-
-            dbg!(&expected_goals, &input_goals);
-            assert_eq!(expected_goals, input_goals);
-        }
+        tasks
     }
 }
